Add Apache 2 license headers
This source code is a merge of two related projects. The Proai caching
provider (Apache licensed) and the Fedora 3 oaiprovider (no license).

Tho provide at least some licensing, the Apache 2 License is applied,
since one of the projects already carries this license and Fedora 3
itself is also licensed under Apache 2.

Note that this should not mean that SLUB claims the copyright for the
individual merged projects.