Add support for multiple set queries
Although the OAI-PMH specification doesn't describe it, this
implementation allows for multiple set-parameters to be present in the
query. All queried setSpecs need to present on a record to match the
query.

The request-Element of the response echoes all queried sets in the
set-Attribute in the form of a comma separated list, e.g.:

<request verb="ListIdentifiers" metadataPrefix="oai_dc"
         set="status-type:publishedVersion, ddc:550">
http://localhost:8081/oai
</request>